Struct flowide::FlowIde

pub(crate) struct FlowIde {
    pub(crate) client: CliRuntimeClient,
    pub(crate) flow_manifest_url: String,
    pub(crate) flow_args: String,
    pub(crate) parallel_jobs_limit: Option<usize>,
}

Fields
client: CliRuntimeClientflow_manifest_url: Stringflow_args: Stringparallel_jobs_limit: Option<usize>
Implementations

impl FlowIde

pub(crate) fn get_lib_search_path( search_path_additions: &[String] ) -> Result<Simpath, Error>

For the lib provider, libraries maybe installed in multiple places in the file system. In order to find the content, a FLOW_LIB_PATH environment variable can be configured with a list of directories in which to look for the library in question.

pub(crate) fn client_and_coordinator( num_threads: usize, lib_search_path: Simpath, native_flowstdlib: bool, debug_this_flow: bool ) -> Result<CliRuntimeClient, Error>

Start a Coordinator in a background thread, then start a client in the calling thread

pub(crate) fn coordinator( num_threads: usize, lib_search_path: Simpath, native_flowstdlib: bool, coordinator_connection: CoordinatorConnection, debug_connection: CoordinatorConnection, loop_forever: bool ) -> Result<(), Error>

Create a new Coordinator, pre-load any libraries in native format that we want to have before loading a flow and it’s library references, then enter the submission_loop() accepting and executing flows submitted for execution, executing each one using the Coordinator

pub(crate) fn client( runtime_client_connection: ClientConnection, debug_this_flow: bool, discovery_post: u16 ) -> Result<CliRuntimeClient, Error>

Create the client that talks to the coordinator

pub(crate) fn flow_url(&self) -> Result<Url>

Create absolute file:// Url for flow location

pub(crate) fn get_flow_args(&self) -> Vec<String>

Create array of strings that are the args to the flow

pub(crate) fn submit(&mut self, debug_this_flow: bool) -> Result<(), Error>

pub(crate) fn get_connect_addresses( ports: (u16, u16, u16, u16) ) -> (String, String, String, String)

Return addresses and ports to be used for each of the three queues

(general) job source

context job source

results sink

control messages

pub(crate) fn get_bind_addresses( ports: (u16, u16, u16, u16) ) -> (String, String, String, String)

Return addresses to bind to for

(general) job source

context job source

results sink

control messages

pub(crate) fn get_four_ports() -> Result<(u16, u16, u16, u16), Error>

Return four free ports to use for client-coordinator message queues

pub(crate) fn num_threads(matches: &ArgMatches) -> usize

Determine the number of threads to use to execute flows

default (if value is not provided on the command line) of the number of cores

pub(crate) fn get_matches() -> ArgMatches

Parse the command line arguments using clap