Module flowide::widgets

Usage

gtk_refs!(
    pub mod widgets;                // The macro emits a new module with this name
    struct WidgetRefs;              // The macro emits a struct with this name containing:
    main_window : gtk::Window ,     // widget_name : Widgettype
    button1 : gtk::Button           // ..
);

fn main() {

    if gtk::init().is_err() {
        println!("Failed to initialize GTK.");
        return;
    }

    let window = Window::new(WindowType::Toplevel);
    window.set_title("gtk-rs-state Example Program");
    window.set_default_size(350, 70);
    let button = Button::new_with_label("Spawn another thread!");
    window.add(&button);
    window.show_all();

    window.connect_delete_event(|_, _| {
        gtk::main_quit();
        Inhibit(false)
    });

    button.connect_clicked(|_| {
        std::thread::spawn(some_workfunction);
        println!("Clicked!");
    });

    // You need the following two statements to prepare the
    // static storage needed for cross thread access.
    // See the `from_glade.rs` example for a more elegant solution
    let widget_references = widgets::WidgetRefs {
        main_window: window.clone(),
        button1:     button.clone(),
    };

    widgets::init_storage(widget_references);
    // End

    // This type has a function for each of your widgets.
    // These functions return a clone() of the widget.
    window.show_all();

    window.connect_delete_event(move |_, _| {
        gtk::main_quit();
        Inhibit(false)
    });

    // Start event loop
    gtk::main();
}

fn compute() {
    use std::thread::sleep;
    use std::time::Duration;
    sleep(Duration::from_secs(1));
}

fn some_workfunction()  {
    let mut i = 0;

    loop {
        compute();

        i += 1;
        let text = format!("Round {} in {:?}", i, std::thread::current().id());

        widgets::do_in_gtk_eventloop(|refs| {
            refs.button1().set_label(&text);
        });
    }
}

Structs

TX
This is the sender part of the above RX channel. All non-gtk threads can put their UiActions-Closures here and wait for completion afterwards.

WidgetRefs

Constants

DATA
This variable is only populated in the gtk thread. The inner struct contains a reference counted pointer to all selected widgets.

RX
This variable is only populated in the gtk thread. The gtk thread receives the boxed UiActions-Closures via this channel. Additionaly, another channels sender is transfered in this channel. This sender is used as signal so the non-gtk thread can wait for completion.

Functions

do_in_gtk_eventloop
Call this from non-gtk threads. The closure allows you to modify the ui.

handle_one_callback_in_gtk_thread

init_storage
Initializes the static storages. You must call this function before calling do_in_gtk_eventloop().

init_storage_from_builder

Type Definitions

BoxedUiAction

FnAndEvent